Change CI, license and parent POM
Also use Circle CI instead of Travis CI as of now. This allows reusing the same
encrypted signing key for all plugins as the secret environment variables are
shared among all ASERVO projects.

Furthermore change the license to Apache 2.0 as agreed upon.

And last but not least, use the latest improvements of the ASERVO parent pom.